share book extension function
